Move prioritizer function EqualPriority to package priorities
